Fix conf centering error.
This commit fixes a bug introduced in the centering of the generalized positions of several robots. Centering is used in some robots to be able to define the group representation in the fields or reals instead of complex numbers.

The issue was related to not properly handling the centering of different joint topologies (revolute, spherical, unbounded prismatic/revolute) etc.