[WIP] Add optional CG_tensor_prod at input of EMLP
There is a design problem with EMLP networks, in which signals from different irrep types (different order/frequency/symmetry pattern) are not mixed within the network.

Essentially meaning that we could think of the EMLP as individual networks for each irrep type. This problem, which becomes prevalent in systems with larger symmetry groups, required some redesign of the network architecture and a careful use of Steerable Wigner Eckart kernel basis, which already handles the CG tensor product.

For the moment we added an argument in the instantiation of the EMLP network `augment_cg_tensor_prod` which, applies a single Clebst-Gordan tensor product at the input of the network, thus mixing the irrep signals in polynomial functions. This tensor product is appended to the "input" of the network.

We can think of this as an additional layer.

This is a temporary solution.